Move new tests to `breakdown_cases`
